backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1

# ── Database (MongoDB Atlas — async) ──────────────────────────
motor==3.4.0
pymongo==4.7.2

# ── Authentication ─────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ── AI / Face Recognition ─────────────────────────────────────
# DeepFace uses FaceNet internally for embedding extraction
deepface==0.0.93
tf-keras==2.16.0
# Fallback: facenet-pytorch for pure PyTorch path (optional)
# facenet-pytorch==2.5.3

# ── Image Processing ──────────────────────────────────────────
Pillow>=10.3.0
opencv-python-headless>=4.9.0
numpy>=1.26.4

# ── YOLO (person/face detection) ──────────────────────────────
ultralytics==8.2.18

# ── Utilities ─────────────────────────────────────────────────
requests>=2.31.0
aiofiles>=23.2.1
scipy>=1.13.0    # cosine distance calculation
